Pass second test
Removed the .skip in the repeatString.spec.js
file.

Added a condition in the repeatString.js
file in the function repeatString where if the
number given is less than zero, then it will
return string 'ERROR'. Also added a loop that
returns the given amount of string that needs to
be repeated given the string and number.