Pass third test
Different from the solution but still passed all
the tests. Need to remember next time to think
about what methods I can possibly use to make
the code less complex.

Solution used split('') to make the string into
an array and then using reverse() to reverse the
array and join('') to convert it back into a
string.

The solution that I came up with was using a for
loop that loops through the length of the array
by decrementing down. Created a string outside
of the loop and added that string with the given
string with method charAt().