Passed test three
Created a different way of implementing the
function compared to the solution. Used a nested
loop, first using a for loop and then used a while
loop.

Was not sure how the function's parameters
worked when given more than what the parameters
asked for. However, I now know that if the
arguments are more than the parameters you set,
then it will make an array like object. Sort of
like a dictionary in python.