Pass test six
My solution used multiple if statements to see
which years are leap years and which are not.

The solution folder showed me that some codes
can be very simple, and simple conditions are all
you need. If statements were not needed and could
have put it all in one return line.